Improves emulator
- Rewrites TMU behavior to match hardware
- Improves output of memory addresses on errors
- Improves/fixes bounds detection for memory access
- Initializes VPM cache with dummy values, not zeroes
- Improves SFU error based on tests on hardware

+++ test/TestMathFunctions.cpp
@@ -544,7 +544,7 @@ void TestMathFunctions::testRint()
 void TestMathFunctions::testRootn()
 {
     testBinaryFunction<16, NormalDistribution<0>>(config, "-DOUT=float16 -DIN0=float16 -DIN1=int16 -DFUNC=rootn",
-        FlushAndRoundBinary{[](float a, int b) -> float { return std::pow(a, 1.0f / b); }},
+        FlushAndRoundBinary{[](float a, int b) -> float { return std::pow(a, 1.0f / static_cast<float>(b)); }},
         std::bind(&TestMathFunctions::onMismatch, this, std::placeholders::_1, std::placeholders::_2));
 }
